requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
boto3==1.35.90
botocore==1.35.90
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2.17
django-bootstrap5==24.3
django-storages==1.14.4
django-widget-tweaks==1.5.0
flatbuffers==24.12.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
idna==3.10
jmespath==1.0.1
keras==3.7.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
pip==24.3.1
protobuf==4.25.5
psycopg2-binary==2.9.10
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
s3transfer==0.10.4
setuptools==65.5.0
six==1.17.0
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0